+++ CHANGES.md
@@ -7,6 +7,9 @@
 * Add support for the `bdi` element
 * Add support for the `search` element
 * Add support for the `data` element
+* Add support for the `track` element and its `kind`, `srclang`
+  (as `a_track_srclang`) and `default` attributes; `audio` and `video`
+  now accept an optional `?tracks` argument
 * Fix typo `whitout` in type definition
 	(#324 by Martin @MBodin Bodin)
 * Add support for the clip-path presentation attribute